third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-invalid-integrity-expected.txt

CONSOLE ERROR: Error parsing 'integrity' attribute ('yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8='). The value must begin with 'ni:///'.
CONSOLE ERROR: Error parsing 'integrity' attribute ('ni:///#yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8='). The specified hash algorithm must be one of 'sha256', 'sha384', or 'sha512'.
CONSOLE ERROR: Error parsing 'integrity' attribute ('yM5ZyzNsyKfaXRY78zSGapeQKtl0oGdpPpYxgwl8XW8='). The value must begin with 'ni:///'.
CONSOLE ERROR: Error parsing 'integrity' attribute ('ni:///sha1;Ywd5te2RvxmuHLzLzZjYo1/0LMw='). The specified hash algorithm must be one of 'sha256', 'sha384', or 'sha512'.
This test passes if no 'FAIL' alert appears.
